OSAL clean-up
