+++ Os/Generic/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Types")
 add_custom_target("${FPRIME_CURRENT_MODULE}")
 
-#### Os/File/Posix Section ####
+#### Os/Generic/Queue Section ####
 set(SOURCE_FILES
         "${CMAKE_CURRENT_LIST_DIR}/PriorityQueue.cpp"
 )
@@ -14,7 +14,6 @@ set(HEADER_FILES
 register_fprime_module(Os_Generic_PriorityQueue)
 register_fprime_implementation(Os/Queue Os_Generic_PriorityQueue "${CMAKE_CURRENT_LIST_DIR}/DefaultPriorityQueue.cpp")
 
-
 set(UT_SOURCE_FILES
         "${CMAKE_CURRENT_LIST_DIR}/test/ut/PriorityQueueTests.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/../test/ut/queue/CommonTests.cpp"
@@ -23,7 +22,6 @@ set(UT_SOURCE_FILES
 set(UT_MOD_DEPS
         Fw_Time
         Os
-        Os_Models
         STest
 )
 choose_fprime_implementation(Os/Queue Os_Generic_PriorityQueue)
