+++ Svc/CmdSequencer/test/ut/CmdSequencerTester.cpp
@@ -13,6 +13,8 @@
 #include "Svc/CmdSequencer/test/ut/SequenceFiles/FPrime/FPrime.hpp"
 #include "CmdSequencerTester.hpp"
 #include "Os/Delegate.hpp"
+#include "Os/Stub/FileSystem.hpp"
+#include "Os/Stub/Directory.hpp"
 
 namespace Svc {
 
@@ -827,4 +829,24 @@ FileInterface *FileInterface::getDelegate(HandleStorage& aligned_placement_new_m
             aligned_placement_new_memory, to_copy
     );
 }
+
+//! \brief get a delegate for FileSystemInterface that intercepts calls for stub fileSystem usage
+//! \param aligned_new_memory: aligned memory to fill
+//! \param to_copy: pointer to copy-constructor input
+//! \return: pointer to delegate
+FileSystemInterface *FileSystemInterface::getDelegate(FileSystemHandleStorage& aligned_placement_new_memory) {
+    return Os::Delegate::makeDelegate<FileSystemInterface, Os::Stub::FileSystem::StubFileSystem>(
+        aligned_placement_new_memory
+    );
+}
+
+//! \brief get a delegate for DirectoryInterface that intercepts calls for stub Directory usage
+//! \param aligned_new_memory: aligned memory to fill
+//! \return: pointer to delegate
+DirectoryInterface *DirectoryInterface::getDelegate(DirectoryHandleStorage& aligned_placement_new_memory) {
+    return Os::Delegate::makeDelegate<DirectoryInterface, Os::Stub::Directory::StubDirectory>(
+        aligned_placement_new_memory
+    );
+}
+
 }
