+++ Svc/PrmDb/test/ut/PrmDbImplTester.cpp
@@ -12,7 +12,8 @@
 #include <Os/Delegate.hpp>
 #include <cstdio>
 #include <gtest/gtest.h>
-
+#include "Os/Stub/FileSystem.hpp"
+#include "Os/Stub/Directory.hpp"
 #include <Fw/Test/UnitTest.hpp>
 
 namespace Svc {
@@ -639,6 +640,25 @@ namespace Os {
 FileInterface *FileInterface::getDelegate(HandleStorage& aligned_placement_new_memory, const FileInterface* to_copy) {
     return Os::Delegate::makeDelegate<FileInterface, Svc::PrmDbImplTester::PrmDbTestFile>(
             aligned_placement_new_memory, to_copy
+    );
+}
+
+//! \brief get a delegate for FileSystemInterface that intercepts calls for stub fileSystem usage
+//! \param aligned_new_memory: aligned memory to fill
+//! \param to_copy: pointer to copy-constructor input
+//! \return: pointer to delegate
+FileSystemInterface *FileSystemInterface::getDelegate(FileSystemHandleStorage& aligned_placement_new_memory) {
+    return Os::Delegate::makeDelegate<FileSystemInterface, Os::Stub::FileSystem::StubFileSystem>(
+        aligned_placement_new_memory
+    );
+}
+
+//! \brief get a delegate for DirectoryInterface that intercepts calls for stub Directory usage
+//! \param aligned_new_memory: aligned memory to fill
+//! \return: pointer to delegate
+DirectoryInterface *DirectoryInterface::getDelegate(DirectoryHandleStorage& aligned_placement_new_memory) {
+    return Os::Delegate::makeDelegate<DirectoryInterface, Os::Stub::Directory::StubDirectory>(
+        aligned_placement_new_memory
     );
 }
 }
