+++ cmake/API.cmake
@@ -615,7 +615,9 @@ function(require_fprime_implementation IMPLEMENTATION)
     endif()
     resolve_dependencies(IMPLEMENTATION "${IMPLEMENTATION}")
     resolve_dependencies(REQUESTER "${REQUESTER}")
-
+    if ("${IMPLEMENTATION}" STREQUAL "Os_FileSystem")
+        message(FATAL_ERROR "Oh Shit")
+    endif ()
     create_implementation_interface("${IMPLEMENTATION}")
     append_list_property("${IMPLEMENTATION}" GLOBAL PROPERTY "REQUIRED_IMPLEMENTATIONS")
     add_dependencies("${REQUESTER}" "${IMPLEMENTATION}")
@@ -635,6 +637,9 @@ endfunction()
 function(register_fprime_implementation IMPLEMENTATION IMPLEMENTOR)
     resolve_dependencies(IMPLEMENTATION "${IMPLEMENTATION}")
     resolve_dependencies(IMPLEMENTOR "${IMPLEMENTOR}")
+    if ("${IMPLEMENTATION}" STREQUAL "Os_FileSystem")
+        message(FATAL_ERROR "Oh Shit")
+    endif ()
     create_implementation_interface("${IMPLEMENTATION}")
     append_list_property("${IMPLEMENTOR}" GLOBAL PROPERTY "${IMPLEMENTATION}_IMPLEMENTORS")
     append_list_property("${ARGN}" TARGET "${IMPLEMENTOR}" PROPERTY "REQUIRED_SOURCE_FILES")
@@ -654,6 +659,9 @@ endfunction()
 function(choose_fprime_implementation IMPLEMENTATION IMPLEMENTOR)
     resolve_dependencies(IMPLEMENTATION "${IMPLEMENTATION}")
     resolve_dependencies(IMPLEMENTOR "${IMPLEMENTOR}")
+    if ("${IMPLEMENTATION}" STREQUAL "Os_FileSystem")
+        message(FATAL_ERROR "Oh Shit")
+    endif ()
     # Check for passed in module name
     if (ARGC EQUAL 3)
         set(ACTIVE_MODULE "${ARGV2}")
