+++ cmake/platform/Darwin.cmake
@@ -19,8 +19,6 @@ if (NOT DEFINED FPRIME_USE_BAREMETAL_SCHEDULER)
    FIND_PACKAGE ( Threads REQUIRED )
 endif()
 choose_fprime_implementation(Os/File Os/File/Posix)
-choose_fprime_implementation(Os/FileSystem Os/FileSystem/Posix)
-choose_fprime_implementation(Os/Directory Os/Directory/Posix)
 choose_fprime_implementation(Os/Console Os/Console/Posix)
 choose_fprime_implementation(Os/Task Os/Task/Posix)
 choose_fprime_implementation(Os/Mutex Os/Mutex/Posix)
